# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BYG 總後台 API",
        "description": "BYG 總後台 API 集成",
        "contact": {
            "email": "asdcxz179@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "測試區"
        }
    ],
    "paths": {
        "/": {},
        "/admin/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "管理員登入",
                "description": "管理員登入",
                "operationId": "8199380c719d0a33f987f7b81154eaf1",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "account",
                                    "password"
                                ],
                                "properties": {
                                    "account": {
                                        "description": "帳號",
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "description": "密碼",
                                        "type": "string",
                                        "example": "123qwe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "success",
                                    "data": {
                                        "id": "1",
                                        "name": "Admin",
                                        "account": "admin",
                                        "token": "54|psaNlF3yDz7iO2cOM80Gun0Par8PWYkCeBlyp6oca60285d4"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "管理員登出",
                "description": "管理員登出",
                "operationId": "21814b2636498c3409222a7acfa9e80f",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "success",
                                    "data": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/users?draw={draw}&start={start}&length={length}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "管理員列表",
                "description": "管理員列表",
                "operationId": "3debe0a2c90c8b0fce648f4e6aa1d1e8",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    },
                    {
                        "name": "draw",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "start",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "length",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "success",
                                    "data": {
                                        "exception": null,
                                        "headers": [],
                                        "original": {
                                            "draw": 1,
                                            "recordsTotal": 1,
                                            "recordsFiltered": 1,
                                            "data": [
                                                {
                                                    "id": 1,
                                                    "name": "admin",
                                                    "account": "admin",
                                                    "status": 1,
                                                    "created_at": "2021-07-01 00:00:00",
                                                    "login_count": 1,
                                                    "last_login_time": "2021-07-01 00:00:00"
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "管理員詳細資料",
                "description": "管理員詳細資料",
                "operationId": "4e24ca9f2b7d5951a6612fe22aa64ed3",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "success",
                                    "data": {
                                        "id": 1,
                                        "name": "admin",
                                        "account": "admin",
                                        "status": 1,
                                        "created_at": "2021-07-01 00:00:00",
                                        "login_count": 1,
                                        "last_login_time": "2021-07-01 00:00:00"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "修改管理員",
                "description": "修改管理員",
                "operationId": "952aa9a9237cb312c9e569bdff3afed1",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "管理員ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "account",
                                    "password",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "名稱",
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "password": {
                                        "description": "密碼",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password_confirmation": {
                                        "description": "確認密碼",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "status": {
                                        "description": "狀態",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "修改成功"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "刪除管理員",
                "description": "刪除管理員",
                "operationId": "5cdf1a1896e99e8ec48144bf7404729d",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "管理員ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "刪除成功"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/users": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "新增管理員",
                "description": "新增管理員",
                "operationId": "8e50211ec84110c16a8b2f98c092b63b",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "account",
                                    "password",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "名稱",
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "account": {
                                        "description": "帳號",
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "password": {
                                        "description": "密碼",
                                        "type": "string",
                                        "example": "123qwe"
                                    },
                                    "password_confirmation": {
                                        "description": "確認密碼",
                                        "type": "string",
                                        "example": "123qwe"
                                    },
                                    "status": {
                                        "description": "狀態",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "新增成功"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "responses": {
            "200": {
                "description": "成功",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 200,
                            "message": "OK"
                        }
                    }
                }
            },
            "400": {
                "description": "客戶端錯誤",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 400,
                            "message": "客戶端錯誤"
                        }
                    }
                }
            },
            "401": {
                "description": "身份驗證失敗",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 401,
                            "message": "Unauthorized"
                        }
                    }
                }
            },
            "404": {
                "description": "找不到請求的資源",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 404,
                            "message": "Not Found"
                        }
                    }
                }
            },
            "405": {
                "description": "不支援此方法",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 405,
                            "message": "Method Not Allowed"
                        }
                    }
                }
            },
            "500": {
                "description": "伺服器發生錯誤",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "status": 500,
                            "message": "伺服器發生錯誤"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Admin",
            "description": "Admin"
        }
    ]
}